shell app refactor
