+++ projects/storefrontapp/src/app/app.module.ts
@@ -10,16 +10,14 @@ import {
 import { StoreDevtoolsModule } from '@ngrx/store-devtools';
 import { translationChunksConfig, translations } from '@spartacus/assets';
 import { ConfigModule, TestConfigModule } from '@spartacus/core';
-import {
-  JsonLdBuilderModule,
-  StorefrontComponent,
-} from '@spartacus/storefront';
-import { b2bFeature } from '../environments/b2b/b2b.feature';
-import { b2cFeature } from '../environments/b2c/b2c.feature';
-import { cdcFeature } from '../environments/cdc/cdc.feature';
-import { cdsFeature } from '../environments/cds/cds.feature';
+import { StorefrontComponent } from '@spartacus/storefront';
 import { environment } from '../environments/environment';
 import { TestOutletModule } from '../test-outlets/test-outlet.module';
+import { HttpClientModule } from '@angular/common/http';
+import { StoreModule } from '@ngrx/store';
+import { EffectsModule } from '@ngrx/effects';
+import { AppRoutingModule } from './app-routing.module';
+import { SpartacusModule } from './spartacus/spartacus.module';
 
 registerLocaleData(localeDe);
 registerLocaleData(localeJa);
@@ -30,27 +28,15 @@ if (!environment.production) {
   devImports.push(StoreDevtoolsModule.instrument());
 }
 
-let additionalImports = [];
-
-if (environment.cds) {
-  additionalImports = [...additionalImports, ...cdsFeature.imports];
-}
-
-if (environment.b2b) {
-  additionalImports = [...additionalImports, ...b2bFeature.imports];
-} else {
-  additionalImports = [...additionalImports, ...b2cFeature.imports];
-}
-
-if (environment.cdc) {
-  additionalImports = [...additionalImports, ...cdcFeature.imports];
-}
-
 @NgModule({
   imports: [
     BrowserModule.withServerTransition({ appId: 'spartacus-app' }),
     BrowserTransferStateModule,
-    JsonLdBuilderModule,
+    HttpClientModule,
+    AppRoutingModule,
+    StoreModule.forRoot({}),
+    EffectsModule.forRoot([]),
+    SpartacusModule,
     ConfigModule.withConfig({
       backend: {
         occ: {
@@ -79,7 +65,6 @@ if (environment.cdc) {
         level: '2.1',
       },
     }),
-    ...additionalImports,
     TestOutletModule, // custom usages of cxOutletRef only for e2e testing
     TestConfigModule.forRoot({ cookie: 'cxConfigE2E' }), // Injects config dynamically from e2e tests. Should be imported after other config modules.
