+++ projects/storefrontapp/src/app/spartacus/features/order-approval-feature.module.ts
@@ -26,4 +26,4 @@ import {
     }),
   ],
 })
-export class OrderApprovalModule {}
+export class OrderApprovalFeatureModule {}
