+++ projects/storefrontapp/src/app/spartacus/features/storefinder-feature.module.ts
@@ -23,4 +23,4 @@ import {
     }),
   ],
 })
-export class StorefinderModule {}
+export class StorefinderFeatureModule {}
