+++ projects/storefrontapp/src/app/spartacus/spartacus-b2c-configuration.module.ts
@@ -34,4 +34,4 @@ import {
     }),
   ],
 })
-export class SpartacusConfigurationModule {}
+export class SpartacusB2cConfigurationModule {}
