--- projects/storefrontapp/src/environments/b2c/app-routing.module.ts
@@ -1,13 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule } from '@angular/router';
-
-@NgModule({
-  imports: [
-    RouterModule.forRoot([], {
-      anchorScrolling: 'enabled',
-      relativeLinkResolution: 'corrected',
-      initialNavigation: 'enabled',
-    }),
-  ],
-})
-export class AppRoutingModule {}
