--- projects/storefrontapp/src/environments/b2c/spartacus/features/storefinder.module.ts
@@ -1,26 +1,0 @@
-import { NgModule } from '@angular/core';
-import { StoreFinderRootModule } from '@spartacus/storefinder/root';
-import { provideConfig } from '@spartacus/core';
-import {
-  storeFinderTranslationChunksConfig,
-  storeFinderTranslations,
-} from '@spartacus/storefinder/assets';
-
-@NgModule({
-  imports: [StoreFinderRootModule],
-  providers: [
-    provideConfig({
-      featureModules: {
-        storeFinder: {
-          module: () =>
-            import('@spartacus/storefinder').then((m) => m.StoreFinderModule),
-        },
-      },
-      i18n: {
-        resources: storeFinderTranslations,
-        chunks: storeFinderTranslationChunksConfig,
-      },
-    }),
-  ],
-})
-export class StorefinderModule {}
