--- projects/storefrontapp/src/environments/b2c/spartacus/spartacus.module.ts
@@ -1,14 +1,0 @@
-import { NgModule } from '@angular/core';
-import { SpartacusConfigurationModule } from './spartacus-configuration.module';
-import { SpartacusFeaturesModule } from './spartacus-features.module';
-import { BaseStorefrontModule } from '@spartacus/storefront';
-
-@NgModule({
-  imports: [
-    BaseStorefrontModule,
-    SpartacusFeaturesModule,
-    SpartacusConfigurationModule,
-  ],
-  exports: [BaseStorefrontModule],
-})
-export class SpartacusModule {}
